refactor: move DetachedDocumentService and model classes to repository package
Renamed DetachedDocumentService to DetachedDocumentRepository and moved
it along with model classes into nl.info.zac.document.detacheddocument.repository.
