+++ src/test/kotlin/nl/info/zac/app/detacheddocuments/converter/RestDetachedDocumentConverterTest.kt
@@ -12,8 +12,8 @@ import io.mockk.every
 import io.mockk.mockk
 import nl.info.zac.app.identity.converter.RestUserConverter
 import nl.info.zac.app.model.createRESTUser
-import nl.info.zac.document.detacheddocument.model.DetachedDocument
-import nl.info.zac.document.detacheddocument.model.createDetachedDocument
+import nl.info.zac.document.detacheddocument.repository.model.DetachedDocument
+import nl.info.zac.document.detacheddocument.repository.model.createDetachedDocument
 import nl.info.zac.enkelvoudiginformatieobject.EnkelvoudigInformatieObjectLockService
 import nl.info.zac.enkelvoudiginformatieobject.model.createEnkelvoudigInformatieObjectLock
 import java.time.LocalDate
