+++ src/test/kotlin/nl/info/zac/app/zaak/ZaakRestServiceCreateTest.kt
@@ -61,7 +61,7 @@ import nl.info.zac.app.zaak.model.createRestZaakCreateData
 import nl.info.zac.authentication.LoggedInUser
 import nl.info.zac.authentication.createLoggedInUser
 import nl.info.zac.configuration.ConfigurationService
-import nl.info.zac.document.detacheddocument.DetachedDocumentService
+import nl.info.zac.document.detacheddocument.repository.DetachedDocumentRepository
 import nl.info.zac.exception.ErrorCode
 import nl.info.zac.flowable.bpmn.BpmnService
 import nl.info.zac.flowable.bpmn.model.createZaaktypeBpmnConfiguration
@@ -101,7 +101,7 @@ class ZaakRestServiceCreateTest : BehaviorSpec({
     val indexingService = mockk<IndexingService>()
     val loggedInUserInstance = mockk<Instance<LoggedInUser>>()
     val objectsClientService = mockk<ObjectsClientService>()
-    val detachedDocumentService = mockk<DetachedDocumentService>()
+    val detachedDocumentService = mockk<DetachedDocumentRepository>()
     val opschortenZaakHelper = mockk<SuspensionZaakHelper>()
     val policyService = mockk<PolicyService>()
     val productaanvraagService = mockk<ProductaanvraagService>()
